Changed gradle structure to use 2 subprojects for deployment
Goal of this was to make deployment possible, as the API implementor and the modloader need to be different jars due to forge restrictions. currently not working as forge on 1.13.3 does not support IModLocators from the mod folder. works on 1.14.4 tho, so we are upgrading for now and maybe later think about going back